Production Release v0.1.2

+++ Modules/HR/Observers/Recruitment/JobObserver.php
@@ -5,6 +5,8 @@ namespace Modules\HR\Observers\Recruitment;
 use Modules\HR\Entities\Job;
 use Modules\HR\Entities\Round;
 use Corcel\Model\Post as Corcel;
+use Corcel\Model\TermRelationship as TermRelationship;
+use Corcel\Model\Term as Term;
 
 class JobObserver
 {
@@ -16,18 +18,28 @@ class JobObserver
      */
     public function created(Job $job)
     {
+        if (! config('database.connections.wordpress.enabled')) {
+            return;
+        }
+
         $job->rounds()->attach(Round::pluck('id')->toArray());
         $data = request()->all();
-
-        $Corcel = new Corcel();
-        $Corcel->post_title = $data['title'];
-        $Corcel->post_content = $data['description'];
-        $Corcel->post_type = config('hr.post-type.career');
+        $corcel = new Corcel();
+        $corcel->post_title = $data['title'];
+        $corcel->post_content = $data['description'];
+        $corcel->post_type = config('hr.post-type.career');
+        $corcel->post_name = str_replace(' ', '-', strtolower($data['title']));
         if ($data['status'] != 'published') {
-            $Corcel->post_status = 'draft';
+            $corcel->post_status = 'draft';
         }
-        $Corcel->save();
-        $Corcel->saveMeta('hr_id', $job['id']);
+        $corcel->save();
+        $corcel->saveMeta('hr_id', $job['id']);
+        $post = Corcel::hasMeta('hr_id', $job['id'])->first();
+        $term = Term::select('term_id')->where(['name' => $data['domain']])->first();
+        $relation = new TermRelationship();
+        $relation->object_id = $post->ID;
+        $relation->term_taxonomy_id = $term->term_id;
+        $relation->save();
     }
 
     /**
@@ -38,15 +50,21 @@ class JobObserver
      */
     public function updated(Job $job)
     {
-        $data = request()->all();
+        if (! config('database.connections.wordpress.enabled')) {
+            return;
+        }
 
+        $data = request()->all();
         $post = Corcel::hasMeta('hr_id', $job['id'])->first();
-        $Corcel = Corcel::find($post->ID);
-        $Corcel->post_title = $data['title'];
-        $Corcel->post_content = $data['description'];
-        $Corcel->post_type = config('hr.post-type.career');
-        $Corcel->post_status = $data['status'] == 'published' ? 'publish' : 'draft';
-        $Corcel->update();
+        $corcel = Corcel::find($post->ID);
+        $corcel->post_title = $data['title'];
+        $corcel->post_content = $data['description'];
+        $corcel->post_type = config('hr.post-type.career');
+        $corcel->post_status = $data['status'] == 'published' ? 'publish' : 'draft';
+        $corcel->post_name = str_replace(' ', '-', strtolower($data['title']));
+        $corcel->update();
+        $term = Term::select('term_id')->where(['name' => $data['domain']])->first();
+        $relation = TermRelationship::where(['object_id' => $post->ID])->update(['term_taxonomy_id' => $term->term_id]);
     }
 
     /**
@@ -57,6 +75,10 @@ class JobObserver
      */
     public function deleted(Job $job)
     {
+        if (! config('database.connections.wordpress.enabled')) {
+            return;
+        }
+
         Corcel::where(['post_type' => 'career', 'post_title' => $job['title']])->delete();
     }
 }
